Data Layer: Introduce response parser
This patch is motivated by two questions:
 - Is the response from the API something I expect?
 - Is the shape of the data in the API respnose what I need?

Previous work has been explored to form a sort of validator function and
an additional `fromAPI()` function to validate and transform data in the
`onSuccess` handlers in the data layer, but that leads to unecessary
code duplication and mess.

The code in this PR is motivated by the extremely capable system of
monadic JSON parsers available in Elm. It's built from the two key
components answering the questions above: using `is-my-json-valid` we
can validate response data and filter out only those fields listed in
the schema; using an optional transformer function we can then reshape
the data once we know it's good. Any failure in this process then leads
to an automatic flow into the `onError()` handler even though the
network request was a 2xx success.